% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentationsR-package.R
\name{CSACSM24}
\alias{CSACSM24}
\title{CSACSM24: Presentation for the 2024 Central States ACSM Regional Meeting}
\description{
Useful functions for accelerometry data processing and analysis and
    a presentation for the 2024 Centeral States ACSM Regional Meeting. The title
    of the presentation is "An introduction to open-source R software for wearable
    device data processing and analysis". The presentation can be downloaded at
    \url{https://github.com/bhelsel/CSACSM24}.
}
\section{CSACSM24 functions}{


\code{\link{render_presentation}}
}

