% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentationsR-package.R
\docType{package}
\name{presentationsR-package}
\alias{presentationsR}
\alias{presentationsR-package}
\title{presentationsR: Presentation for the 2024 Central States ACSM Regional Meeting}
\description{
Presentation for the 2024 Central States ACSM Regional Meeting. This presentation is titled An introduction to open-source R software for wearable device data processing and analysis.
}
\keyword{internal}
