% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{render_presentation}
\alias{render_presentation}
\title{render_presentation}
\usage{
render_presentation(name, ...)
}
\arguments{
\item{name}{Name of the presentation or workshop to render}

\item{...}{Additional parameters passed to quarto::quarto_render}
}
\value{
A rendered Quarto revealjs presentation in the output directory
    that is opened on the user's default Internet browser
}
\description{
Copies presentation to an output directory and opens the HTML presentation in web browser
}
\details{
Copies presentation to an output directory and opens the HTML presentation in web browser
}
\seealso{
\code{\link[quarto]{quarto_render}}
 \code{\link[utils]{browseURL}}
}
